1，删除测试类 2，重新编译so库，编译所有指令集的库 3，修复long、int类型在64位和32位编译期间发生的兼容bug

+++ amrcodec/src/main/java/cn/wzh/amrcodec/decoder/AmrDecoder.java
@@ -11,7 +11,7 @@ public class AmrDecoder {
         System.loadLibrary("amr-codec");
     }
 
-    public static native int initDecamr();
+    public static native long initDecamr();
 
     public static native void exitDecAmr(int decode);
